gemma-slm-hosting/requirements.txt
httpx
numpy
python-dotenv
requests
tqdm
